requirements.txt
gradio
opencv-python
torch
torchvision
scipy
matplotlib
ffmpeg-python
shutil
pickle
threading
os
multiprocessing
pillow
argparse
tqdm